Refactoring board_generation package into subpackages
Refactoring into subpackages


Update imports in benchmarking
