+++ ic_routing_board_generation/board_generator/numpy_board_generation/board_generator_wfc_oj.py
@@ -1,9 +1,7 @@
-import random
-from typing import List, Tuple
 import numpy as np
 from copy import deepcopy
 
-from ic_routing_board_generation.board_generator.abstract_board import AbstractBoard
+from ic_routing_board_generation.board_generator.numpy_data_model.abstract_board import AbstractBoard
 
 class AbstractTile():
     """
