+++ ic_routing_board_generation/ic_rl_training/ic_generator.py
@@ -13,20 +13,16 @@
 # limitations under the License.
 
 import abc
-from typing import Optional
 
 import chex
 import jax
 import jax.numpy as jnp
 
-from ic_routing_board_generation.board_generator.utils import \
-    get_heads_and_targets
-from ic_routing_board_generation.interface.board_generator_interface_jax import \
-    BoardGenerator, BoardName
+from ic_routing_board_generation.board_generator.jax_board_generation.board_generator_random_seed_rb import \
+    RandomSeedBoard
 from jumanji.environments.routing.connector.types import Agent, State
 from jumanji.environments.routing.connector.utils import get_position, get_target
 
-from ic_routing_board_generation.board_generator.most_basic_rw import JaxRandomWalk
 
 class Generator(abc.ABC):
     """Base class for generators for the connector environment."""
@@ -58,7 +54,175 @@ class Generator(abc.ABC):
         """
 
 
-class UniformRandomGenerator(Generator):
+# class UniformRandomGenerator(Generator):
+#     """Randomly generates `Connector` grids that may or may not be solvable. This generator places
+#     start and target positions uniformly at random on the grid.
+#     """
+#
+#     def __init__(self, grid_size: int, num_agents: int) -> None:
+#         """Instantiates a `UniformRandomGenerator`.
+#
+#         Args:
+#             grid_size: size of the square grid to generate.
+#             num_agents: number of agents/paths on the grid.
+#         """
+#         super().__init__(grid_size, num_agents)
+#
+#     def __call__(self, key: chex.PRNGKey) -> State:
+#         """Generates a `Connector` state that contains the grid and the agents' layout.
+#
+#         Returns:
+#             A `Connector` state.
+#         """
+#         key, pos_key = jax.random.split(key)
+#         starts_flat, targets_flat = jax.random.choice(
+#             key=pos_key,
+#             a=jnp.arange(self.grid_size**2),
+#             shape=(2, self.num_agents),  # Start and target positions for all agents
+#             replace=False,  # Start and target positions cannot overlap
+#         )
+#
+#         # Create 2D points from the flat arrays.
+#         starts = jnp.divmod(starts_flat, self.grid_size)
+#         targets = jnp.divmod(targets_flat, self.grid_size)
+#
+#         # Get the agent values for starts and positions.
+#         agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
+#         agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
+#
+#         # Create empty grid.
+#         grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
+#
+#         # Place the agent values at starts and targets.
+#         grid = grid.at[starts].set(agent_position_values)
+#         grid = grid.at[targets].set(agent_target_values)
+#
+#         # Create the agent pytree that corresponds to the grid.
+#         agents = jax.vmap(Agent)(
+#             id=jnp.arange(self.num_agents),
+#             start=jnp.stack(starts, axis=1),
+#             target=jnp.stack(targets, axis=1),
+#             position=jnp.stack(starts, axis=1),
+#         )
+#
+#         step_count = jnp.array(0, jnp.int32)
+#
+#         return State(key=key, grid=grid, step_count=step_count, agents=agents)
+
+
+# class ICGenerators(Generator):
+#     """Randomly generates `Connector` grids that may or may not be solvable. This generator places
+#     start and target positions uniformly at random on the grid.
+#     """
+#
+#     def __init__(self, grid_size: int, num_agents: int, board_generator: Optional[BoardName] = None) -> None:
+#         """Instantiates a `UniformRandomGenerator`.
+#
+#         Args:
+#             grid_size: size of the square grid to generate.
+#             num_agents: number of agents/paths on the grid.
+#         """
+#         super().__init__(grid_size, num_agents)
+#         self.board_generator = board_generator
+#
+#     def __call__(self, key: chex.PRNGKey) -> State:
+#         """Generates a `Connector` state that contains the grid and the agents' layout.
+#
+#         Returns:
+#             A `Connector` state.
+#         """
+#
+#         if self.board_generator == "uniform":
+#             key, pos_key = jax.random.split(key)
+#             starts_flat, targets_flat = jax.random.choice(
+#                 key=pos_key,
+#                 a=jnp.arange(self.grid_size ** 2),
+#                 shape=(2, self.num_agents),
+#                 # Start and target positions for all agents
+#                 replace=False,  # Start and target positions cannot overlap
+#             )
+#
+#             # Create 2D points from the flat arrays.
+#             starts = jnp.divmod(starts_flat, self.grid_size)
+#             targets = jnp.divmod(targets_flat, self.grid_size)
+#
+#             # Get the agent values for starts and positions.
+#             agent_position_values = jax.vmap(get_position)(
+#                 jnp.arange(self.num_agents))
+#             agent_target_values = jax.vmap(get_target)(
+#                 jnp.arange(self.num_agents))
+#
+#             # Create empty grid.
+#             grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
+#
+#             # Place the agent values at starts and targets.
+#             grid = grid.at[starts].set(agent_position_values)
+#             grid = grid.at[targets].set(agent_target_values)
+#
+#             # Create the agent pytree that corresponds to the grid.
+#             agents = jax.vmap(Agent)(
+#                 id=jnp.arange(self.num_agents),
+#                 start=jnp.stack(starts, axis=1),
+#                 target=jnp.stack(targets, axis=1),
+#                 position=jnp.stack(starts, axis=1),
+#             )
+#
+#             step_count = jnp.array(0, jnp.int32)
+#
+#         else:
+#             key, pos_key = jax.random.split(key)
+#             starts_flat, targets_flat = jax.random.choice(
+#                 key=pos_key,
+#                 a=jnp.arange(self.grid_size**2),
+#                 shape=(2, self.num_agents),  # Start and target positions for all agents
+#                 replace=False,  # Start and target positions cannot overlap
+#             )
+#
+#             # Create 2D points from the flat arrays.
+#             starts = jnp.divmod(starts_flat, self.grid_size)
+#             targets = jnp.divmod(targets_flat, self.grid_size)
+#
+#             # Get the agent values for starts and positions.
+#             agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
+#             agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
+#
+#             # Create empty grid.
+#             grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
+#
+#             board_class = BoardGenerator.get_board_generator(
+#                 board_enum=self.board_generator)
+#             board = board_class(self.grid_size, self.grid_size, self.num_agents)
+#             pins = board.return_training_board()
+#             grid = jnp.array(pins, int)
+#
+#             starts_flat, targets_flat = get_heads_and_targets(pins)
+#
+#             starts = jnp.divmod(starts_flat, self.grid_size)
+#             targets = jnp.divmod(targets_flat, self.grid_size)
+#
+#             agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
+#             agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
+#
+#             # Place the agent values at starts and targets.
+#             grid = grid.at[starts].set(agent_position_values)
+#             grid = grid.at[targets].set(agent_target_values)
+#
+#             # Create the agent pytree that corresponds to the grid.
+#
+#             agents = jax.vmap(Agent)(
+#                 id=jnp.arange(self.num_agents),
+#                 start=jnp.stack(starts, axis=1),
+#                 target=jnp.stack(targets, axis=1),
+#
+#                 position=jnp.stack(starts, axis=1),
+#             )
+#
+#             step_count = jnp.array(0, jnp.int32)
+#
+#         return State(key=key, grid=grid, step_count=step_count, agents=agents)
+
+
+class RandomSeedGenerator(Generator):
     """Randomly generates `Connector` grids that may or may not be solvable. This generator places
     start and target positions uniformly at random on the grid.
     """
@@ -71,6 +235,7 @@ class UniformRandomGenerator(Generator):
             num_agents: number of agents/paths on the grid.
         """
         super().__init__(grid_size, num_agents)
+        self.board_generator = RandomSeedBoard(grid_size, grid_size, num_agents)
 
     def __call__(self, key: chex.PRNGKey) -> State:
         """Generates a `Connector` state that contains the grid and the agents' layout.
@@ -79,193 +244,21 @@ class UniformRandomGenerator(Generator):
             A `Connector` state.
         """
         key, pos_key = jax.random.split(key)
-        starts_flat, targets_flat = jax.random.choice(
-            key=pos_key,
-            a=jnp.arange(self.grid_size**2),
-            shape=(2, self.num_agents),  # Start and target positions for all agents
-            replace=False,  # Start and target positions cannot overlap
-        )
 
-        # Create 2D points from the flat arrays.
-        starts = jnp.divmod(starts_flat, self.grid_size)
-        targets = jnp.divmod(targets_flat, self.grid_size)
-
-        jax.debug.print("starts are {x}", x=starts)
-        jax.debug.print("targets are {x}", x=targets)
-
-        # Get the agent values for starts and positions.
-        agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
-        agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
-
-        # Create empty grid.
         grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
-
-        # Place the agent values at starts and targets.
-        grid = grid.at[starts].set(agent_position_values)
-        grid = grid.at[targets].set(agent_target_values)
-
-        # Create the agent pytree that corresponds to the grid.
-        agents = jax.vmap(Agent)(
-            id=jnp.arange(self.num_agents),
-            start=jnp.stack(starts, axis=1),
-            target=jnp.stack(targets, axis=1),
-            position=jnp.stack(starts, axis=1),
-        )
-
-        step_count = jnp.array(0, jnp.int32)
-
-        return State(key=key, grid=grid, step_count=step_count, agents=agents)
-
-
-class ICGenerators(Generator):
-    """Randomly generates `Connector` grids that may or may not be solvable. This generator places
-    start and target positions uniformly at random on the grid.
-    """
-
-    def __init__(self, grid_size: int, num_agents: int, board_generator: Optional[BoardName] = None) -> None:
-        """Instantiates a `UniformRandomGenerator`.
-
-        Args:
-            grid_size: size of the square grid to generate.
-            num_agents: number of agents/paths on the grid.
-        """
-        super().__init__(grid_size, num_agents)
-        self.board_generator = board_generator
-
-    def __call__(self, key: chex.PRNGKey) -> State:
-        """Generates a `Connector` state that contains the grid and the agents' layout.
-
-        Returns:
-            A `Connector` state.
-        """
-
-        if self.board_generator == "uniform":
-            key, pos_key = jax.random.split(key)
-            starts_flat, targets_flat = jax.random.choice(
-                key=pos_key,
-                a=jnp.arange(self.grid_size ** 2),
-                shape=(2, self.num_agents),
-                # Start and target positions for all agents
-                replace=False,  # Start and target positions cannot overlap
-            )
-
-            # Create 2D points from the flat arrays.
-            starts = jnp.divmod(starts_flat, self.grid_size)
-            targets = jnp.divmod(targets_flat, self.grid_size)
-
-            # Get the agent values for starts and positions.
-            agent_position_values = jax.vmap(get_position)(
-                jnp.arange(self.num_agents))
-            agent_target_values = jax.vmap(get_target)(
-                jnp.arange(self.num_agents))
-
-            # Create empty grid.
-            grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
-
-            # Place the agent values at starts and targets.
-            grid = grid.at[starts].set(agent_position_values)
-            grid = grid.at[targets].set(agent_target_values)
-
-            # Create the agent pytree that corresponds to the grid.
-            agents = jax.vmap(Agent)(
-                id=jnp.arange(self.num_agents),
-                start=jnp.stack(starts, axis=1),
-                target=jnp.stack(targets, axis=1),
-                position=jnp.stack(starts, axis=1),
-            )
-
-            step_count = jnp.array(0, jnp.int32)
-
-        else:
-            key, pos_key = jax.random.split(key)
-            starts_flat, targets_flat = jax.random.choice(
-                key=pos_key,
-                a=jnp.arange(self.grid_size**2),
-                shape=(2, self.num_agents),  # Start and target positions for all agents
-                replace=False,  # Start and target positions cannot overlap
-            )
-
-            # Create 2D points from the flat arrays.
-            starts = jnp.divmod(starts_flat, self.grid_size)
-            targets = jnp.divmod(targets_flat, self.grid_size)
-
-            # Get the agent values for starts and positions.
-            agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
-            agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
-
-            # Create empty grid.
-            grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
-
-            board_class = BoardGenerator.get_board_generator(
-                board_enum=self.board_generator)
-            board = board_class(self.grid_size, self.grid_size, self.num_agents)
-            pins = board.return_training_board(key)
-            grid = jnp.array(pins, int)
-
-            starts_flat, targets_flat = get_heads_and_targets(pins)
-
-            starts = jnp.divmod(starts_flat, self.grid_size)
-            targets = jnp.divmod(targets_flat, self.grid_size)
-
-            agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
-            agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
-
-            # Place the agent values at starts and targets.
-            grid = grid.at[starts].set(agent_position_values)
-            grid = grid.at[targets].set(agent_target_values)
-
-            # Create the agent pytree that corresponds to the grid.
-
-            agents = jax.vmap(Agent)(
-                id=jnp.arange(self.num_agents),
-                start=jnp.stack(starts, axis=1),
-                target=jnp.stack(targets, axis=1),
-
-                position=jnp.stack(starts, axis=1),
-            )
-
-            step_count = jnp.array(0, jnp.int32)
-
-        return State(key=key, grid=grid, step_count=step_count, agents=agents)
-
-
-class JaxRandomWalkGenerator(Generator):
-    """Randomly generates `Connector` grids that may or may not be solvable. This generator places
-    start and target positions uniformly at random on the grid.
-    """
-
-    def __init__(self, grid_size: int, num_agents: int) -> None:
-        """Instantiates a `UniformRandomGenerator`.
-
-        Args:
-            grid_size: size of the square grid to generate.
-            num_agents: number of agents/paths on the grid.
-        """
-        super().__init__(grid_size, num_agents)
-
-    def __call__(self, key: chex.PRNGKey) -> State:
-        """Generates a `Connector` state that contains the grid and the agents' layout.
-
-        Returns:
-            A `Connector` state.
-        """
-        key, pos_key = jax.random.split(key)
-        board_generator = JaxRandomWalk(self.grid_size, self.grid_size, self.num_agents)
-        
-        grid = jnp.zeros((self.grid_size, self.grid_size), dtype=jnp.int32)
-        starts, targets = board_generator.generate_starts_ends(key)
+        starts, targets = self.board_generator.generate_starts_ends(key)
         # jax.debug.print("our starts: {x}", x=starts)
         # jax.debug.print("our targets: {x}", x=targets)
-        agent_position_values = jax.vmap(get_position)(jnp.arange(self.num_agents))
+        agent_position_values = jax.vmap(get_position)(
+            jnp.arange(self.num_agents))
         agent_target_values = jax.vmap(get_target)(jnp.arange(self.num_agents))
         # jax.debug.print("agent_position_values: {x}", x=agent_position_values)
         # Transpose the agent_position_values to match the shape of the grid.
         # Place the agent values at starts and targets.
         grid = grid.at[starts].set(agent_position_values)
         grid = grid.at[targets].set(agent_target_values)
-        
+
         # Create the agent pytree that corresponds to the grid.
-        
 
         agents = jax.vmap(Agent)(
             id=jnp.arange(self.num_agents),
@@ -278,13 +271,3 @@ class JaxRandomWalkGenerator(Generator):
         step_count = jnp.array(0, jnp.int32)
 
         return State(key=key, grid=grid, step_count=step_count, agents=agents)
-
-
-if __name__ == "__main__":
-    test = UniformRandomGenerator(10, 5)
-    key = jax.random.PRNGKey(0)
-    test(key)
-
-    test = JaxRandomWalkGenerator(10, 5)
-    key = jax.random.PRNGKey(0)
-    test(key)
