+++ ic_routing_board_generation/ic_rl_training/ic_setup_train.py
@@ -25,10 +25,8 @@ import optax
 from omegaconf import DictConfig
 
 from ic_routing_board_generation.ic_rl_training.ic_generator import \
-    ICGenerators, UniformRandomGenerator
-from ic_routing_board_generation.ic_rl_training.dataset_generator import \
-    BoardDatasetGenerator
-from ic_routing_board_generation.interface.board_generator_interface_jax import \
+     RandomSeedGenerator #ICGenerators, UniformRandomGenerator,
+from ic_routing_board_generation.interface.board_generator_interface import \
     BoardName
 from jumanji.env import Environment
 from jumanji.environments import (
@@ -148,23 +146,22 @@ def setup_logger(cfg: DictConfig) -> Logger:
 def _make_raw_env(cfg: DictConfig, ic_generator: Optional[BoardName] = None) -> Environment:
     # env: Environment = jumanji.make(cfg.env.registered_version)
     # TODO (RL): remove ic_generator parameter and return to jumanji.make, remove imports
-    if cfg.env.ic_board.generation_type == "offline":
-        print("offline")
-        generator = BoardDatasetGenerator(
-            grid_size=cfg.env.ic_board.grid_size,
-            num_agents=cfg.env.ic_board.num_agents,
-            board_generator=cfg.env.ic_board.board_name,
-        )
-    elif cfg.env.ic_board.generation_type == "online":
-        print("online")
-        generator = ICGenerators(
-            grid_size=cfg.env.ic_board.grid_size,
-            num_agents=cfg.env.ic_board.num_agents,
-            board_generator=cfg.env.ic_board.board_name,
-        )
-    else:
-        raise ValueError("please enter valid generation type online or offline")
+    # TODO (MW): add generator
 
+    generator = RandomSeedGenerator(
+        grid_size=cfg.env.ic_board.grid_size,
+        num_agents=cfg.env.ic_board.num_agents,
+    )
+    # generator = ICGenerators(
+    #     grid_size=cfg.env.ic_board.grid_size,
+    #     num_agents=cfg.env.ic_board.num_agents,
+    #     board_generator=cfg.env.ic_board.board_name,
+    # )
+    # generator = UniformRandomGenerator(
+    #     grid_size=cfg.env.ic_board.grid_size,
+    #     num_agents=cfg.env.ic_board.num_agents,
+    #     # board_generator=cfg.env.ic_board.board_name,
+    # )
     env = Connector(generator=generator)
     if isinstance(env, Connector):
         env = MultiToSingleWrapper(env)
