+++ ic_routing_board_generation/visualisation/render_rough.py
@@ -1,14 +1,13 @@
 # Render boards from env_viewer.py
 import os
-import time
 from typing import Optional, Union
 import numpy as np
-from ic_routing_board_generation.board_generator.abstract_board import AbstractBoard
-from ic_routing_board_generation.board_generator.bfs_board import BFSBoard
-from ic_routing_board_generation.board_generator.board_generator_numberlink_oj import NumberLinkBoard
-from ic_routing_board_generation.board_generator.board_generator_random_walk_rb import RandomWalkBoard
-from ic_routing_board_generation.board_generator.board_generator_wfc_oj import WFCBoard
-from ic_routing_board_generation.board_generator.board_processor_2 import BoardProcessor, BFS_fill, LSystem_fill
+from ic_routing_board_generation.board_generator.numpy_data_model.abstract_board import AbstractBoard
+from ic_routing_board_generation.board_generator.numpy_board_generation.bfs_board import BFSBoard
+from ic_routing_board_generation.board_generator.numpy_board_generation.board_generator_numberlink_oj import NumberLinkBoard
+from ic_routing_board_generation.board_generator.numpy_board_generation.board_generator_random_walk_rb import RandomWalkBoard
+from ic_routing_board_generation.board_generator.numpy_board_generation.board_generator_wfc_oj import WFCBoard
+from ic_routing_board_generation.board_generator.numpy_utils.board_processor_2 import BoardProcessor, BFS_fill, LSystem_fill
 from ic_routing_board_generation.board_generator.lsystem_board import LSystemBoardGen
 from ic_routing_board_generation.visualisation.env_viewer import RoutingViewer
